fix lst

+++ ft_strtrim.c
@@ -30,9 +30,9 @@ static int	is_set(char c, char const *set)
 
 char		*ft_strtrim(char const *s1, char const *set)
 {
-	size_t	i;
-	size_t	j;
-	size_t	k;
+	int		i;
+	int		j;
+	int		k;
 	char	*str;
 
 	if (!s1)
